pyproject.toml
[project]
name = "certificate-api"
version = "1.0.0"
description = "FastAPI server for certificate generation and email delivery"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "email-validator>=2.0.0",
    "sqlmodel>=0.0.21",
    "pymysql>=1.1.0",
    "sqlacodegen>=4.0.0",
    "alembic>=1.13.0",
]

# to run the server
# USE: fastapi dev
